+++ ProjectE-Integration/src/api/java/com/tagnumelite/projecteintegration/api/utils/ConfigHelper.java
@@ -28,7 +28,6 @@ import net.minecraftforge.common.config.Configuration;
 public final class ConfigHelper {
     public static final String CATEGORY_GENERAL = "general";
     public static final String CATEGORY_PLUGINS = "plugins";
-    public static final String CATEGORY_MISC = "misc";
 
     public static String getCategory(String... txts) {
         return String.join(Configuration.CATEGORY_SPLITTER, txts);
@@ -42,14 +41,6 @@ public final class ConfigHelper {
         return getCategory(CATEGORY_PLUGINS, modid);
     }
 
-    public static String getResourceCategory() {
-        return getCategory(CATEGORY_GENERAL, "resources");
-    }
-
-    public static String resourceIntoCFGName(ResourceLocation resource) {
-        return resource.getResourceDomain() + '-' + resource.getResourcePath();
-    }
-
     public static String getConversionName(String name) {
         return "enable_" + ConfigHelper.getConfigName(name) + "_conversions";
     }
@@ -58,7 +49,11 @@ public final class ConfigHelper {
         return txt.toLowerCase().replaceAll("[ \\-~`\\[{}\\]+='\";:/?.>,<!@#$%^&*()]", "_");
     }
 
+    public static String getMapperName(String name) {
+        return "enable_" + getConfigName(name) + "_mapper";
+    }
+
     public static String getMapperName(PEIMapper mapper) {
-        return "enable_" + getConfigName(mapper.name) + "_mapper";
+        return getMapperName(mapper.name);
     }
 }